Split resolver-backed helper tests
